Start adding bgr conversion code

+++ src/main/kotlin/com/github/ajsnarr98/hauntedgameboard/hardware/camera/RealLibCamera.kt
@@ -47,7 +47,7 @@ class RealLibCamera : Camera {
     }
 
     class RawPicture {
-        var pixels: IntArray = IntArray(0) // expected BGR format
+        var pixels: ByteArray = ByteArray(0) // expected BGR format
         var width: Int = 0
         var height: Int = 0
     }
